Updates to build system focused on auto-gen'd repo targets.
Since eval'ing a directive make determining the contents of the
vars within the directive difficult just start eval'ing the vars
which forces immediate resolution and expansion.

This stuff is a massive PITA.  eval'ing targets doesn't have this
issue.  Just variables.

Also some other minor cleanups.

The build system can now populate the pkglist files and
populate and manage the repos in repos/.

There is still a bug in the mock conf for RHEL 6.  It is still
looking for a buildsys group that it can't find.  This issue is
preventing RPMs from rolling in mock.