Remainder of pungi mods to gen isos.
I meant to commit the orig pungi first and the patched 2nd but I mixed
it up.

Anyhow here are the original, patched, and the diff.

Right now it gens ISOs that have resolved symlinks in them.  The
ISO looks like it will work correctly but I'm on 3G and I'm
not downloading the ISO to complete the test.