Add support for generating RPMs that must be installed on the build host.
We need pungi installed on the build host.   Pungi isn't in RHEL, EPEL,
or Opt.  But we don't want it added to the CONFIG_BUILD packages list
as it will confuse users and will be error prone.  I added support for
rolling RPMs for the host. It is extendable as usual but the list is
in the Makefile not the CONFIG_BUILD file.

Note: I will add some documentation regarding this and general guidance
for the build host requirements next week.