Add openscap 0.8.5 to build system.
We were carrying the sources for 0.8.4 temporarily.  The last commit
dropped that source.  This commit adds 0.8.5 as a src RPM since we
aren't patching it.  Note this is not the official src RPM.  It was
rolled by Petr Lautrbach to get it out without waiting for it to be
merged into epel.