Fix dep lists in top-level Makefile.
I added a check a little while back that ensures the proper RPMs have
been installed on the host.  This target had an empty deps list which
was causing everything to be re-rolled.  Convert it into a macro and
call it everywhere the old dep existed *within* the relevant target.