Add user_u to SELinux to fix issue with useradd.
useradd is busted.  It runs genhomedircon prior to creating the login
mapping.  What this means is that a policy that lacks the __default__
user mapping, specific mapping (eg bob->staff_u), and user_u it thinks the user home dir context should be
default_t.  This occurs even if you do useradd -Z because it runs
genhomedircon *then* adds the mapping.  The work-around is to leave
user_u in the policy, useradd -Z, run genhomedircon, then run restorecon
-R /home.