Drop CONFIG_BUILD as dep for pkglists.
CONFIG_BUILD was a dep for conf/pkglist*.  This meant changing
PACKAGES etc was trigger a re-creation of the pkglists.  This
was causing problems when devs didn't expect this behavior.  Shift
RHEL_VER out of that file (I'm not planning on rolling RHEL5 from this
repo) so we can drop it as a dep.