Introduce support for blacklisting packages.
We have run into some problems with certain package versions.  We need
to ensure certain package versions are not used.  This commit adds
support for a "blacklist" concept.  You add a package name to
conf/pkglist.blacklist and it is added to yum/mock conf exclude lines.

For now blacklist policycoreutils release 21.