CLIP: Switch to host's patched livecd-creator.
We are now carrying our own livecd-creator due to the need to patch some
components and the expectation we will need to do more patching moving
forward.  Drop the calls to the one we carried in support/.

CXS-15

The modifications in this commit are copyrighted works of the Cubic Corporation.
(C) 2013 Cubic Corporation