CLIP: Add custom dracut module to fix livecd booting w/ SELinux.
The SELinux bits in dracut weren't written for RAM-backed stateless
environments like those found in LiveCDs.  Specifically, if a relabel is
required, it was performed in dracut, but the autorelabel file was never
removed in the system was never put back into enforcing (if necessary).
This is a problem because if autorelabel is present when rc.sysinit runs
the system relabels (again) and reboots.  Since this is stateless all
changes are lost, and the entire thing starts over again, and again, and
again.  Additionally, it wasn't putting the system back into enforcing
mode after a relabel.

This module runs after the default selinux module, removes the
autorelabel file, and sets the enforcing mode appropriately (permissive
or enforcing depending on config).

CXS-15

The modifications in this commit are copyrighted works of the Cubic Corporation.
(C) 2013 Cubic Corporation