CLIP: policy fixes for kernel_t in livecd environments.
In LiveCD environments, where dracut is still running as kernel_t, some
operations are happening and they are failing as the kernel doesn't have
appropriate perms.  Two things, the kernel is trusted, dracut is highly
trusted as it loads the policy itself and could pretty much do whatever
it wants.  While not ideal, until dracut is fixed and either re-execs
itself, or the load_policy occurs after pivot_root, we are stuck with
dracut running in kernel_t.

So, give it permission to read things and relabel things.

CXS-15

The modifications in this commit are copyrighted works of the Cubic Corporation.
(C) 2013 Cubic Corporation