CLIP: add CLIP dracut module to ks.
We have our own pretty little dracut module.  Install it in the ks.
This fixes the livecd, ram-backed file system issues.

Also add it to the CONFIG_BUILD so we can actually build the package.

CXS-15

The modifications in this commit are copyrighted works of the Cubic Corporation.
(C) 2013 Cubic Corporation