CLIP: Use system's livecd-iso-to-disk.
In RHEL 5 we had to modify livecd-iso-to-disk but in RHEL 6 we can use
the system's (plus ours no longer worked).

Switch to the system's livecd-iso-to-disk and drop the old one in our
tree.

THis commit also drops livecd-creator from our support dir which we are no longer using.

CXS-15

The modifications in this commit are copyrighted works of the Cubic Corporation.
(C) 2013 Cubic Corporation