CLIP: Fix 32-bit builds.
Hadn't tested on i386 in awhile. A mock template file for i386
was not present.  The bootstrap script used a hard-coded arch
in one place.

Add mock template file for i386.
Use host arch instead of hard-coded arch in bootstrap.

The modifications in this commit are copyrighted works of the Cubic Corporation.
(C) 2013 Cubic Corporation

CXS-20