Copy all generated RPMs to OUTPUT_DIR outside of mock.
RPMs can have subpackages.  Copy all RPMs to the OUTPUT_DIR
instead of just the single RPM with the package name. Also,
some arch-specific packages may generate noarch RPMs so
copy those too.  This part just uses globbing which is
fine for creating RPMs, but removing them is a different story.

The most complicated piece of this was getting an exhaustive list
of packages for bare targets.  This commit adds support for
extracting package and subpackage info from spec files and uses
that for bare targets.

Issue #21