Add device-mapper package.
Live media on an ISO will fail to boot
without dmsetup, and will appear to
build successfully.

This fixes that problem by adding the package.
In reality, the 90dmsquash-live dracut module's
check should ensure dmsetup did indeed get
installed in the RD.  But, I'm not even sure if
these checks will cause a failure in livecd-creator.