irqbalance has a tight loop on getsched
irqbalance spins in a tight loop on getsched().
When it is denied in enforcing mode this causes
shit to hit the fan.  It generates 100,000s of AVCs
which the auditd actually consumes but is failing to
flush to disk fast enough so auditd just keeps
consuming more and more memory as irqbalance spins.
Eventually OOM killer runs and decides irqbalance needs
to die.  Anyhow, this should fix all of that.