Extract version and release from spec files in src.rpm.
It turns out that the packages generated by some
src.rpms don't match the src.rpm's version/release. E.g.,
the src.rpm is 1.0-123rf, and it generates 1.0-123el6.

This adds support to the scripts we use to generate
Makefiles for src.rpm packages to extract the spec file
and query that for version/release instaed of the src.rpm.

This problem actually poped up b/c a package wlways built
due to dep issues.  make thought a different release
was going to be generated than actually was.