More srpm generation fixes for arch detection.
The top-level makes some assumptions regarding
how to construct and src.rpm name.  This leads
to problems when a src.pm ver/rel differs
from the pakacge ver/rel.  This splits the
concepts of a src.rpm into two files for packages
rolled directly from src.rpms.

Additionally, extrac an arch from the spec file
in case it is noarch.