Issue #37: Add more optional packages needed by lorax
Added at-spi, tigervnc-server-module, and bitmap-fangsongti-fonts to
bootstrap. These packages will no longer spew out an error during a
build because they were missing. There is still and error for
*-redhat-upgrade-dracut, but this is documented in
https://bugzilla.redhat.com/show_bug.cgi?format=multiple&id=1065557.
Since this package is not yet available, but will soon be added, it
should not be removed.

Removed rpm-ostree and kexec-tools-anaconda-addon from the lorax templates.
We don't use these packages and it silences some errors during the build.

Run createrepo on the opt repo each time we bootstrap. This needs to be
done because a new RPM might be added to the OPT_PACKAGES list sometime
in the future.