Issue #106: Create a cache directory for lorax to use
A cache directory will now be created in
tmp/clip-build-iso/work/x86_64/cache/. After a build, installroot will be
copied to the cache directory. This will ensure that RPMs will not have
to be reinstalled when rebuilding.

Build times are reduced from 10 mins to 4 mins on the Tresys build
system.

To ensure a fresh build is created without using the cache, run 'make clean'.