Issue #171: Fix TAR_DEPS variable in clip-selinux-policy
The clip-selinux-policy Makefile used the version of the package
when collecting files to track to build the tarball for the src rpm.
Since the clip-selinux-policy source directory does not have a version,
the files were not gathered properly. This allows mock to function correctly
and detect code changes that will trigger rebuilding the RPM when
necessary.