Correct ARCH determination of in-tree packages
The PKG_ARCH Makefile macro to get the architecture of a package
correctly greps and evaluates ARCH from the packages Makefile, but then
incorrectly returns TARGET_ARCH instead of ARCH. TARGET_ARCH is always
the architecture of the build (e.g. x86_64) which can be different from
the architecture of a package if it is a noarch package. This difference
leads to make unnecessarily rebuilding rpms because it expects an
x86_64.rpm file but that package will create a noarch.rpm file. This
patch changes this to correctly return the ARCH of the package.