Fix issue of random incorrect ARCH
The main Makefile greps and evaluates the ARCH line from each packages
Makefile to determine the ARCH for packages. Some ARCH lines use ?= and
some use := to set this variable. But depending on the order of when
make evaluates ARCH, if ?= is used then the evaluation might be ignored
if ARCH was previously evaluated for another package. This can cause
incorrect behaviors.

Instead, make sure that we always use := when setting an ARCH so that
the eval always overrides a previous ARCH. To allow for ARCH to be set
to something other than the current architecture (e.g. to cross-compile)
the TARGET_ARCH variable can be set.