Fix and simplify NTP lockdown
The sed script had an error in that it used &1 instead of \1 for a
backrefernce. This resulted in it adding a "1" to the NTP configuration
which causes NTP to fail to start.

This simplifies the sed scripts so they are more clear, and makes sure
that no stray 1's are added.