Remove pam_lastlog from pam system-auth/password-auth
RHEL7 already includes the "postlogin" configuration file in the
appropriate pam files for displaying last login information when a user
logs in. This is necessary to meet SSG requirements.

However, CLIP also adds pam_lastlog in the system-auth/password-auth
files. Not only does this causes duplicate last login information to be
displayed, it also causes last login information to be displayed at
times when it shouldn't, like when someone runs a command via sudo.

This removes pam_lastlog from system-auth and password-auth, relying on
postlogin to be used and to do the right thing.