Updates for latest refpolicy
usbguard needs to be in the base due to a conditional

fix for refpolicy commit 2b5d1b829976db11a56fada424032ffeff46ef7b
In this commit, fstab generated was labeled with a new type and
granted access to write init_rumtime_t, but we had it labeled
systemd_unit_t with this change.  This part of this patch can now
go away and things work again.